Rename method names

+++ azur-engine/src/CircleColliderComponent.cpp
@@ -21,11 +21,11 @@ void CircleColliderComponent::draw(SDL_Renderer* renderer)
 	else SDL_SetRenderDrawColor(renderer, 0, 255, 0, 255); // green
 
 	// DEBUG: Draw circle for testing
-	int32_t centerX = position->x;
-	int32_t centerY = position->y;
-	const int32_t diameter = (radius * 2);
+	int32_t centerX = (int32_t)position->x;
+	int32_t centerY = (int32_t)position->y;
+	const int32_t diameter = int32_t(radius * 2.0f);
 
-	int32_t x = (radius - 1);
+	int32_t x = int32_t(radius - 1);
 	int32_t y = 0;
 	int32_t tx = 1;
 	int32_t ty = 1;
@@ -33,14 +33,14 @@ void CircleColliderComponent::draw(SDL_Renderer* renderer)
 
 	while (x >= y)
 	{
-		SDL_RenderDrawPoint(renderer, position->x + x, position->y - y);
-		SDL_RenderDrawPoint(renderer, position->x + x, position->y + y);
-		SDL_RenderDrawPoint(renderer, position->x - x, position->y - y);
-		SDL_RenderDrawPoint(renderer, position->x - x, position->y + y);
-		SDL_RenderDrawPoint(renderer, position->x + y, position->y - x);
-		SDL_RenderDrawPoint(renderer, position->x + y, position->y + x);
-		SDL_RenderDrawPoint(renderer, position->x - y, position->y - x);
-		SDL_RenderDrawPoint(renderer, position->x - y, position->y + x);
+		SDL_RenderDrawPoint(renderer, int(position->x + x), int(position->y - y));
+		SDL_RenderDrawPoint(renderer, int(position->x + x), int(position->y + y));
+		SDL_RenderDrawPoint(renderer, int(position->x - x), int(position->y - y));
+		SDL_RenderDrawPoint(renderer, int(position->x - x), int(position->y + y));
+		SDL_RenderDrawPoint(renderer, int(position->x + y), int(position->y - x));
+		SDL_RenderDrawPoint(renderer, int(position->x + y), int(position->y + x));
+		SDL_RenderDrawPoint(renderer, int(position->x - y), int(position->y - x));
+		SDL_RenderDrawPoint(renderer, int(position->x - y), int(position->y + x));
 		if (error <= 0)
 		{
 			++y;
